Split class implementations from header files
