+++ Makefile
@@ -1,5 +1,6 @@
 main: main.cpp
-	g++ -std=c++11 main.cpp -o scheduler
+	g++ -std=c++11 main.cpp scheduler.cpp sys.cpp \
+	    -o scheduler
 
 .PHONY: clean
 clean:
