+++ process.h
@@ -8,6 +8,8 @@
 #ifndef PROCESS_H
 #define PROCESS_H	
 
+#include <string>
+
 using namespace std;
 
 struct Process
