+++ scheduler.cpp
@@ -1,30 +1,4 @@
-///////////////////////////////////////////////////////////////////////////////
-// Scheduler.h
-// @desc: Scheduler class header file
-///////////////////////////////////////////////////////////////////////////////
-
-#pragma once
-
-#ifndef SCHEDULER_H
-#define SCHEDULER_H	
-
-#include "Process.h"
-#include <vector>
-#include <climits>
-
-class Scheduler
-{
-public:
-	vector <Process*> mProcesses; // processes to be scheduled
-
-	void clean();	// remove all finished processes from Scheduler
-
-	void FCFS(const int quantum, const int time);	// First-Come, First-Serve scheduling 
-	void RR(const int quantum, const int time);		// Round Robin scheduling 
-	void SJN(const int quantum, const int time);	// Shortest Job Next scheduling 
-	void SRT(const int quantum, const int time);	// Shortest Remaining Time scheduling 
-	void HRRN(const int quantum, const int time);	// Highest Response Ratio Next scheduling 
-};
+#include "scheduler.h"
 
 // TODO: rework logic to increment by time slice instead of quantum
 // Shortest Remaining Time
@@ -76,5 +50,3 @@ void Scheduler::clean()
 			mProcesses.erase(mProcesses.begin()+i);
 	}
 }
-
-#endif
