+++ sys.cpp
@@ -1,31 +1,4 @@
-///////////////////////////////////////////////////////////////////////////////
-// Sys.h
-// @desc: Sys class header file
-///////////////////////////////////////////////////////////////////////////////
-
-#pragma once
-
-#include "Scheduler.h"
-
-#ifndef SYS_H
-#define SYS_H
-
-class Sys
-{
-public:
-	Scheduler* mScheduler;
-	vector <Process> mProcesses;	// process job list read in at start 
-	
-	int mQuantum;	// time units for process execution before rescheduling 
-	int mSchedSel;	// scheduling algorithm selection
-	int mTime;		// current system time
-
-	Sys() { mTime = 0; } // initiliaze run time = 0
-
-	void schedule();		// add newly arrived processes from job list to Scheduler
-	bool jobsFinished();	// return true if all processes in job list are finished
-	void printProgress();	// print progress of all processes on Scheduler
-};
+#include "sys.h"
 
 // add newly arrived processes from job list to Scheduler
 // return: NONE
@@ -86,5 +59,3 @@ void Sys::printProgress()
 		cout << "(" << mScheduler->mProcesses[i]->mProgressT << ")\n";
 	}
 }
-
-#endif
